refactor: error.js파일 삭제 -> post.error.js
