--- src/error.js
@@ -1,9 +1,0 @@
-export class alreadyExistPostLike extends Error {
-    errorCode = "P001";
-  
-    constructor(reason) {
-      super(reason);
-      this.reason = reason;
-      this.statusCode = 400;
-    }
-  }
